mrc-kv-cache-configurator-core(kvc-extract-mla-dim-move,mechanical_provable): Move calculate_mla_kv_cache_dim to mem_cache.kv_cache_configurator (cut+paste)
